Collapse the harvest-to-yield derivation into one place
The derivation existed twice: once on the component selection screen when
a harvest was entered, and once per year in the results pipeline. The two
used different formulas, so they disagreed whenever the bale and crop
moistures differed, and the component screen's copy wrote the harvest onto
the crop itself - from where it reached every year of the stand even for a
harvest marked as belonging to one year only.

CropViewItem.CalculateYieldFromHayHarvests is now the single
implementation. It sums the hayed harvests for its own year, converts to
dry matter, re-expresses that on the crop's moisture basis, and sets the
total biomass, dry yield and yield together. It reports whether it could,
and refuses only when the numbers cannot support a derivation.

UpdateYieldFromHarvestForCustomPerennials keeps the policy - which yield
assignment method, perennial, not grazed, not frozen - and calls it. The
component screen no longer derives a yield at all; see the companion
change in the views repository.
