+++ H.Core/Models/LandManagement/Fields/CropViewItem.Harvest.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Linq;
+using H.Core.Enumerations;
 using H.Core.Models.Results;
 
 namespace H.Core.Models.LandManagement.Fields
@@ -128,7 +129,82 @@ namespace H.Core.Models.LandManagement.Fields
         {
             return this.GetHayHarvests().Any();
         }
-        
+
+        /// <summary>
+        /// Sets this item's yield from the hay actually baled off it, and reports whether it could.
+        ///
+        /// This is the single place the derivation lives. It used to be done in two: once on the component selection
+        /// screen when a harvest was entered, and once per year in the results pipeline - with different formulas, so
+        /// the two disagreed whenever the bale and crop moistures differed.
+        ///
+        /// The two figures are on different moisture bases and have to be reconciled: a bale has been dried (around 15%
+        /// moisture) while Yield means the standing crop as it grew (around 80% for forage), and C_p multiplies Yield by
+        /// 1 - the CROP's moisture. Using the bale's wet weight directly would strip ~80% of the mass off material
+        /// holding only ~15% water. So the harvest's dry matter is re-expressed on the crop's basis, leaving
+        /// Yield x (1 - crop moisture) equal to the dry matter baled off, as the pipeline expects.
+        ///
+        /// Callers own the policy about WHEN to derive (yield assignment method, grazing, frozen values); this method
+        /// only refuses when the numbers cannot support a derivation.
+        /// </summary>
+        public bool CalculateYieldFromHayHarvests()
+        {
+            var hayedHarvests = this.GetHayHarvests()
+                .Where(harvest => harvest.ForageActivity == ForageActivities.Hayed)
+                .ToList();
+
+            if (hayedHarvests.Any() == false || this.Area <= 0)
+            {
+                return false;
+            }
+
+            var totalDryMatter = hayedHarvests.Sum(GetHarvestDryMatter);
+            if (totalDryMatter <= 0)
+            {
+                return false;
+            }
+
+            // A crop recorded as all water leaves nothing to scale by.
+            var dryMatterFraction = 1.0 - this.GetMoistureContentOfCropAsFraction();
+            if (dryMatterFraction <= 0)
+            {
+                return false;
+            }
+
+            // The total biomass stays the wet weight actually baled - it is what the harvest tab reports - while the
+            // yield is the standing-crop equivalent of that biomass.
+            this.TotalBiomassHarvest = hayedHarvests.Sum(harvest => harvest.AboveGroundBiomass);
+            this.DryYield = totalDryMatter / this.Area;
+            this.Yield = this.DryYield / dryMatterFraction;
+
+            return true;
+        }
+
+        /// <summary>
+        /// The dry matter baled off in a harvest. <see cref="FieldActivityBase.AboveGroundBiomassDryWeight"/> is
+        /// maintained whenever the bale count, bale weight or moisture changes, but is recomputed here from the wet
+        /// weight when it was never populated (an older saved farm, or an item built without going through those setters).
+        /// </summary>
+        private static double GetHarvestDryMatter(HarvestViewItem harvestViewItem)
+        {
+            if (harvestViewItem.AboveGroundBiomassDryWeight > 0)
+            {
+                return harvestViewItem.AboveGroundBiomassDryWeight;
+            }
+
+            return harvestViewItem.AboveGroundBiomass * (1.0 - (harvestViewItem.MoistureContentAsPercentage / 100.0));
+        }
+
+        /// <summary>
+        /// The crop's moisture content as a fraction. Some saved farms hold it as a percentage instead, which the carbon
+        /// calculator corrects when it runs; this can be called earlier, so it has to tolerate both forms.
+        /// </summary>
+        public double GetMoistureContentOfCropAsFraction()
+        {
+            var moistureContent = this.MoistureContentOfCrop;
+
+            return moistureContent > 1 ? moistureContent / 100.0 : moistureContent;
+        }
+
         #endregion
     }
 }
